[CWS] extract hostname related utils functions to their own package
