+++ pkg/eventmonitor/consumers/testutil/testutil.go
@@ -16,7 +16,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/eventmonitor"
 	"github.com/DataDog/datadog-agent/pkg/eventmonitor/consumers"
 	eventtestutil "github.com/DataDog/datadog-agent/pkg/eventmonitor/testutil"
-	"github.com/DataDog/datadog-agent/pkg/security/utils"
+	"github.com/DataDog/datadog-agent/pkg/security/utils/hostnameutils"
 )
 
 const defaultChanSize = 100
@@ -26,7 +26,7 @@ const defaultChanSize = 100
 func NewTestProcessConsumer(t *testing.T) *consumers.ProcessConsumer {
 	var pc *consumers.ProcessConsumer
 	// Set fake hostname to avoid fetching it from the core agent.
-	utils.SetCachedHostname("test-hostname")
+	hostnameutils.SetCachedHostname("test-hostname")
 	eventtestutil.StartEventMonitor(t, func(t *testing.T, evm *eventmonitor.EventMonitor) {
 		var err error
 		eventTypes := []consumers.ProcessConsumerEventTypes{consumers.ExecEventType, consumers.ExitEventType}
