+++ pkg/security/utils/hostnameutils/hostname.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package utils holds utils related files
-package utils
+// Package hostnameutils holds utils/hostname related files
+package hostnameutils
 
 import (
 	"context"
